Feature1/Image Classification/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
pydantic>=2.0.0
